Separate logic a bit and move it to another classes
